Modified the layout from the Odin Project project.
I created a new layout from scratch, using three different
stylesheets. That was actually pretty cumbersome, and it might have
some issues with performance, but I don't know if that's the case
or not.

Also added some fonts from google-fonts, as well as some non-
standard layout types so it's not so samey from the other Odin
project projects.